+++ src/cpp/fastrtps_deprecated/types/TypeObject.cpp
@@ -8942,7 +8942,7 @@ size_t to_size_t(const OctetSeq& s)
     // result = s[0] + (s[1] * 255) + (s[2] * 255²) + (s[3] * 255³);
     for (size_t i = 0; i < s.size() && i < 4; ++i)
     {
-        result += (std::pow(255, i) * s[i]);
+        result += (static_cast<size_t>(std::pow(255, i)) * s[i]);
     }
 
     return result;
